refactor: decompose ApiRequestHandler into composable requestPipeline modules
Extract ~565 lines from ApiRequestHandler.ts (815 -> 250 lines) into
independently testable modules under src/core/requestPipeline/:

- headers.ts: buildRequestHeaders, parseCacheControlTtl
- cachePolicy.ts: lookupSpecTtl, trySpecAwareCacheHit, tryStaleCacheResponse, cacheResponse
- statusHandling.ts: handleEarlyStatus, handleErrorResponse, wrapError, STATUS_MESSAGES
- paginationOrchestration.ts: handleCursorPagination, handleOffsetPagination
- middlewareBridge.ts: applyRequestMiddleware, applyResponseInterceptors
- fetchExecution.ts: executeSingleFetch, fetchOnePage, parseJsonBody
- dependencies.ts: resolveCache, resolveRateLimiter, resolveCircuitBreaker, resolveRetryStrategy

ApiRequestHandler.ts is now a thin coordinator containing only executeRequest
and the two public entry points (handleRequest, handleSinglePageRequest).

All existing tests pass unmodified. New focused unit tests added for headers,
cachePolicy, and statusHandling modules (42 new tests).
